+++ server.js
@@ -1,120 +1,114 @@
+#!/usr/bin/env node
+
+const httpsPort = 18554;
+const {LNSERVICE_LND_DIR} = process.env;
+const {join} = require('path');
+const {readFileSync} = require('fs');
+const WebSocketServer = require('ws').Server;
+const {verifyClient} = require('./push');
 const {log} = console;
 
-const asyncAuto = require('async/auto');
-const asyncRetry = require('async/retry');
-const basicAuth = require('express-basic-auth');
-const bodyParser = require('body-parser');
-const compress = require('compression')();
-const config = require('dotenv').config();
-const cors = require('cors');
-const express = require('express');
-const logger = require('morgan');
-const walnut = require('walnut');
+/**
+ * Module dependencies.
+ */
 
-const {addressesRouter} = require('./routers');
-const {authorizer} = require('./routers');
-const {balanceRouter} = require('./routers');
-const {channelsRouter} = require('./routers');
-const {connectionsRouter} = require('./routers');
-const {cryptoRouter} = require('./routers');
-const {exchangeRouter} = require('./routers');
-const {getWalletInfo} = require('./lightning');
-const {historyRouter} = require('./routers');
-const {invoicesRouter} = require('./routers');
-const {isWalletLocked} = require('./service');
-const {lightningDaemon} = require('./lightning');
-const {localLnd} = require('./service');
-const {networkInfoRouter} = require('./routers');
-const {paymentsRouter} = require('./routers');
-const {peersRouter} = require('./routers');
-const {purchasedRouter} = require('./routers');
-const {rowTypes} = require('./lightning');
-const {subscribeToInvoices} = require('./push');
-const {subscribeToTransactions} = require('./push');
-const {transactionsRouter} = require('./routers');
-const {unlockWallet} = require('./lightning');
-const {walletInfoRouter} = require('./routers');
-const {walletPasswordPrompt} = require('./service');
+const app = require('./app');
+const debug = require('debug')('adapter:server');
+const http = require('http');
+const https = require('https');
 
-const app = express();
-const logFormat = ':method :url :status - :response-time ms - :user-agent';
-const {NODE_ENV} = process.env;
-const unlockDelayMs = 1000;
-const unlockerLnd = localLnd({is_unlocker: true});
+/**
+ * Get port from environment and store in Express.
+ */
 
-if (NODE_ENV !== 'production') {
-  walnut.check(require('./package'));
-}
+const port = normalizePort(process.env.PORT || '10553');
+app.set('port', port);
 
-module.exports = app;
+/**
+ * Create HTTP server.
+ */
+const server = app
+  .listen(port, () => log(null, `Listening HTTP on port: ${port}`))
+  .on('error', err => log([500, 'ListenError', err]));
 
-return asyncAuto({
-  // Determine if the wallet is locked
-  isLocked: cbk => isWalletLocked({}, cbk),
+const [cert, key] = ['cert', 'key']
+  .map(extension => join(LNSERVICE_LND_DIR, `tls.${extension}`))
+  .map(n => readFileSync(n, 'utf8'));
 
-  // Get the wallet unlock password
-  walletPassword: ['isLocked', async ({isLocked}, cbk) => {
-    return !isLocked ? Promise.resolve() : await walletPasswordPrompt({});
-  }],
+const httpsServer = https.createServer({cert, key}, app);
 
-  // Unlock the wallet if necessary
-  unlockWallet: [
-    'isLocked',
-    'walletPassword',
-    ({isLocked, walletPassword}, cbk) =>
-  {
-    // Exit early when there is no need to unlock
-    if (!isLocked) {
-      return cbk();
-    }
+httpsServer.listen(httpsPort);
 
-    return unlockWallet({lnd: unlockerLnd, password: walletPassword}, err => {
-      if (!!err) {
-        return cbk(err);
-      }
+const wss = [
+  new WebSocketServer({server, verifyClient}),
+  new WebSocketServer({server: httpsServer, verifyClient})
+];
 
-      return setTimeout(() => cbk(), unlockDelayMs);
-    });
-  }],
+/**
+ * Listen on provided port, on all network interfaces.
+ */
 
-  // Start API server
-  startServer: ['unlockWallet', ({}, cbk) => {
-    const wss = [];
-    const lnd = localLnd({});
+server.listen(port);
+server.on('error', onError);
+server.on('listening', onListening);
 
-    app.disable('x-powered-by');
-    app.use(compress);
-    app.use(cors());
-    app.use(bodyParser.json());
-    app.use(logger(logFormat));
-    app.use(basicAuth({authorizer, authorizeAsync: true}));
+// });
+/**
+ * Normalize a port into a number, string, or false.
+ */
 
-    app.use('/v0/addresses', addressesRouter({lnd, log}));
-    app.use('/v0/balance', balanceRouter({lnd, log}));
-    app.use('/v0/channels', channelsRouter({lnd, log}));
-    app.use('/v0/connections', connectionsRouter({lnd, log}));
-    app.use('/v0/crypto', cryptoRouter({lnd, log}));
-    app.use('/v0/exchange', exchangeRouter({log}));
-    app.use('/v0/history', historyRouter({lnd, log}));
-    app.use('/v0/invoices', invoicesRouter({lnd, log, wss}));
-    app.use('/v0/network_info', networkInfoRouter({lnd, log}));
-    app.use('/v0/payments', paymentsRouter({lnd, log, wss}));
-    app.use('/v0/peers', peersRouter({lnd, log}));
-    app.use('/v0/purchased', purchasedRouter({lnd, log}));
-    app.use('/v0/transactions', transactionsRouter({lnd, log, wss}));
-    app.use('/v0/wallet_info', walletInfoRouter({lnd, log}));
+function normalizePort(val) {
+  const port = parseInt(val, 10);
 
-    subscribeToInvoices({lnd, log, wss});
-    subscribeToTransactions({lnd, log, wss});
-
-    app.wss = wss;
-    return cbk();
-  }],
-},
-err => {
-  if (!!err) {
-    return log(err);
+  if (isNaN(port)) {
+    // named pipe
+    return val;
   }
 
-  return;
-});
+  if (port >= 0) {
+    // port number
+    return port;
+  }
+
+  return false;
+}
+
+/**
+ * Event listener for HTTP server "error" event.
+ */
+
+function onError(error) {
+  if (error.syscall !== 'listen') {
+    throw error;
+  }
+
+  const bind = typeof port === 'string'
+    ? 'Pipe ' + port
+    : 'Port ' + port;
+
+  // handle specific listen errors with friendly messages
+  switch (error.code) {
+    case 'EACCES':
+      console.error(bind + ' requires elevated privileges');
+      process.exit(1);
+      break;
+    case 'EADDRINUSE':
+      console.error(bind + ' is already in use');
+      process.exit(1);
+      break;
+    default:
+      throw error;
+  }
+}
+
+/**
+ * Event listener for HTTP server "listening" event.
+ */
+
+function onListening() {
+  const addr = server.address();
+  const bind = typeof addr === 'string'
+    ? 'pipe ' + addr
+    : 'port ' + addr.port;
+  debug('Listening on ' + bind);
+}
